factor out tf methods form TEDP
